+++ src/Directives/ForInValueOnlyStatement.php
@@ -6,7 +6,7 @@ namespace Caprice\Directives;
  * This file is a part of Caprice package
  *
  * @package     Caprice
- * @version     0.3.0
+ * @version     0.4.0
  * @author      Lotfio Lakehal <contact@lotfio.net>
  * @copyright   Lotfio Lakehal 2019
  * @license     MIT
@@ -23,7 +23,7 @@ class ForInValueOnlyStatement implements DirectiveInterface
      *
      * @var string
      */
-    public $pattern = '/#for\s*\((\$\w+\s*)(\s*in\s*)(.*?)\)(.*?)#endfor/s';
+    public $pattern = '/#for\s*\((\$\w+\s*)(\s*in\s*)([^\r\n]+)\)(.*?)#endfor/s';
 
     /**
      * directive replace method.